Moves JacksonPreStatement as an inner class of StatementImpl
